+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `RoomReactionSender` to client platform.
 
+- Added methods for fetching events to client platform.
+
 ## [3.1.1] - 2025-03-29
 
 ### Added